refatorações e mais

+++ src/main/java/br/com/trabalhopoo/mybar/repository/TipoItemRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.trabalhopoo.mybar.model.TipoItem;
 
-public interface TipoItemRepository  extends JpaRepository<TipoItem, Integer>{
+public interface TipoItemRepository  extends JpaRepository<TipoItem, Long>{
     List<TipoItem> findByAtivoTrue();
 }
